+++ vscode/extension/CHANGELOG.md
@@ -7,6 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## 1.9.0
 
+- Use CodeLens to show dependents instead of inlay hints.
+- Show dependents of .liquid, .md, and .txt files.
 - Lower file edit debounce after saving from 1s to 10ms.
 - Preview: Fallback to rendering with Configurable Strings disabled if there is an error loading.